Restructuring of the package
